# docs/swagger.yaml
basePath: /v2
definitions:
  domain.VolumeStatus:
    properties:
      error_code:
        type: string
      error_message:
        type: string
      status_value:
        type: integer
    type: object
...
  view.ClusterLunCreateRequest:
    properties:
      cluster_id:
        type: integer
      desc:
        type: string
      fs_type:
        type: string
      name:
        type: string
      node_ids:
        type: string
      path:
        type: string
      pr_support_status:
        type: string
      sector_num:
        type: integer
      sector_size:
        type: integer
      size:
        type: integer
      vendor:
        type: string
      wwid:
        type: string
    type: object
  view.ClusterLunFormatRequest:
    properties:
      fs_size:
        type: integer
      fs_type:
        type: string
      name:
        type: string
      wwid:
        type: string
    type: object
  view.ClusterLunFsExpandRequest:
    properties:
      fs_size:
        type: integer
      fs_type:
        type: string
      name:
        type: string
      pv_name:
        type: string
      wwid:
        type: string
    type: object
  view.ClusterLunResponse:
    properties:
      cluster_id:
        type: integer
      db_cluster_name:
        type: string
      desc:
        type: string
      fs_size:
        type: integer
      fs_type:
        type: string
      lv_name:
        type: string
      name:
        type: string
      node_ids:
        type: string
      path_num:
        type: integer
      paths:
        items:
          type: string
        type: array
      pr_support_status:
        type: string
      product:
        type: string
      pvc_name:
        type: string
      sector_num:
        type: integer
      sector_size:
        type: integer
      size:
        type: integer
      status:
        $ref: '#/definitions/domain.VolumeStatus'
      usable:
        type: boolean
      used_size:
        type: integer
      vendor:
        type: string
      volume_type:
        type: string
      wwid:
        type: string
    type: object
...
  view.DmCreateCmdResponse:
    properties:
      lv_name:
        type: string
      lv_type:
        type: string
      preview_conf:
        type: string
      sector_size:
        type: integer
      sectors:
        type: integer
      size:
        type: integer
    type: object
  view.ErrorResult:
    properties:
      error:
        type: string
    type: object
  view.GenerateDmCreateCmdRequest:
    properties:
      lun_ids_in_order:
        items:
          type: string
        type: array
      lv_name:
        type: string
      lv_type:
        type: string
    type: object
...
  view.WorkflowIdResponse:
    properties:
      workflow_id:
        type: string
    type: object
...
host: 127.0.0.1:2002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: v1.8 存储管理系统
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: PolarDB Storage Manage System
  version: "1.0"
paths:
...
  /cluster-luns:
    get:
      consumes:
      - application/json
      description: 用于查询所有 Cluster LUN
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/view.ClusterLunResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 列表 Cluster LUN
      tags:
      - LUN 管理
    post:
      consumes:
      - application/json
      description: 用于创建 Cluster LUN
      parameters:
      - description: 请求参数
        in: body
        name: clusterLun
        required: true
        schema:
          $ref: '#/definitions/view.ClusterLunCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 创建 Cluster LUN
      tags:
      - LUN 管理
  /cluster-luns/:wwid:
    delete:
      consumes:
      - application/json
      description: 用于删除指定 Cluster LUN
      parameters:
      - description: LUN WWID
        in: query
        name: wwid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 删除 Cluster LUN
      tags:
      - LUN 管理
    get:
      consumes:
      - application/json
      description: 用于查询指定 Cluster LUN
      parameters:
      - description: LUN WWID
        in: query
        name: wwid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.ClusterLunResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 查询 Cluster LUN
      tags:
      - LUN 管理
  /cluster-luns/expand:
    post:
      consumes:
      - application/json
      description: 用于扩容指定 Cluster LUN
      parameters:
      - description: 请求参数
        in: body
        name: clusterLun
        required: true
        schema:
          $ref: '#/definitions/view.ClusterLunCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 扩容 Cluster LUN
      tags:
      - LUN 管理
  /cluster-luns/format:
    post:
      consumes:
      - application/json
      description: 用于格式化指定 Cluster LUN
      parameters:
      - description: 请求参数
        in: body
        name: clusterLun
        required: true
        schema:
          $ref: '#/definitions/view.ClusterLunFormatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 格式化 Cluster LUN
      tags:
      - LUN 管理
  /cluster-luns/fs-expand:
    post:
      consumes:
      - application/json
      description: 用于扩容指定 Cluster LUN Filesystem
      parameters:
      - description: 请求参数
        in: body
        name: clusterLun
        required: true
        schema:
          $ref: '#/definitions/view.ClusterLunFsExpandRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 扩容 Cluster LUN For Filesystem
      tags:
      - LUN 管理
...
  /cluster-lvs/dm-conf:
    post:
      consumes:
      - application/json
      description: 用于生成预览的lv的配置文件
      parameters:
      - description: 请求参数
        in: body
        name: clusterLv
        required: true
        schema:
          $ref: '#/definitions/view.GenerateDmCreateCmdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.DmCreateCmdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 生成预览的lv的配置文件
      tags:
      - LV 管理
...
  /events:
    post:
      consumes:
      - application/json
      description: 用于存储事件上报
      parameters:
      - in: query
        name: body
        type: string
      - in: query
        name: eType
        type: integer
      - in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: Event Upload
      tags:
      - Event 接入管理
  /events/batch:
    post:
      consumes:
      - application/json
      description: 用于存储事件上报
      parameters:
      - in: query
        name: node_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: Batch Event Upload
      tags:
      - Event 接入管理
...
  /scsi/rescan:
    post:
      consumes:
      - application/json
      description: 用于扫描 Volume Rescan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/view.WorkflowIdResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/view.ErrorResult'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/view.ErrorResult'
      summary: 扫描 Volume Rescan
      tags:
      - Scsi 管理
...
swagger: "2.0"